Convert data class

+++ app/src/main/java/me/skripsi/rekomendasibeliapp/utils/DataUjiChangedState.kt
@@ -3,5 +3,5 @@ package me.skripsi.rekomendasibeliapp.utils
 sealed class DataUjiChangedState {
     data class Stok(val updatedValue : Int): DataUjiChangedState()
     data class Diskon(val updatedValue : Boolean): DataUjiChangedState()
-    data class Penjualan(val updatedValue : Double): DataUjiChangedState()
+    data class Penjualan(val updatedValue : Int): DataUjiChangedState()
 }
