+++ roomDb/src/main/java/me/skripsi/roomdb/entity/DataTransaksiEntity.kt
@@ -11,6 +11,6 @@ data class DataTransaksiEntity(
     val golongan: String? = "",
     val stok: Int = 0,
     val isDiskon: Short = 0,
-    val penjualan: Double = 0.0,
-    val pembelian: Double = 0.0
+    val penjualan: Int = 0,
+    val pembelian: Int = 0
 )
