+++ roomDb/src/main/java/me/skripsi/roomdb/entity/DataUjiEntity.kt
@@ -15,5 +15,5 @@ data class DataUjiEntity(
     val golongan: String? = "",
     val stok: Int = 0,
     val isDiskon: Boolean = false,
-    val penjualan: Double = 0.0,
+    val penjualan: Int = 0,
 )
